requirements.txt
streamlit
streamlit-folium
folium
requests
pandas
openpyxl
python-dotenv
httpx
beautifulsoup4
